fix:recuperado função de detalhe de filmes

+++ apps/filmes/views.py
@@ -25,7 +25,26 @@ class FilmesViewSet(viewsets.ViewSet):
 
         return Response(resultado)
     
+    def retrieve(self,request,pk=None):
+        if pk is None:
+            return Response({"erro": "ID obrigatório"}, status=400)
     
+        with connection.cursor() as cursor:
+            cursor.execute("""
+                     SELECT *
+                     FROM filmes
+                        WHERE id = %s
+                        """, [pk])
+
+            columns = [col[0] for col in cursor.description]
+            row = cursor.fetchone()
+
+        if not row:
+            return Response({"erro": "Filme não encontrado"}, status=404)
+
+        resultado = dict(zip(columns, row))
+
+        return Response(resultado)
     
     def create(self,request):
         serializer = FilmesAddSerializer(data=request.data)
@@ -80,12 +99,15 @@ class FilmesViewSet(viewsets.ViewSet):
                          }, status=201)        
                 
     def destroy(self,request,pk=None):
+        if pk is None:
+            return Response({"erro": "ID obrigatório"}, status=400)
+        
         with connection.cursor() as cursor:
             cursor.execute("""
                            DELETE FROM filmes
                            WHERE id=%s
                            """,[pk])
-        return Response({"Mensagem:Filme removido com sucesso"},status=204)         
+        return Response({"Mensagem":"Filme removido com sucesso"},status=204)         
     
 class GenerosViewSet(viewsets.ViewSet):
